+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'sara-arch-v239';
+const CACHE_NAME = 'sara-arch-v240';
 const STATIC_ASSETS = [
   '/Sara-Arch/',
   '/Sara-Arch/index.html',